revamped marketplace
this might break stuff but navbar works
