+++ front-end/src/Components/MarketPlace/AllPages.js
@@ -1,0 +1,14 @@
+import { CampaignComp } from './CampaignComp';
+import { SolutionComp } from './SolutionComp';
+import { ServiceComp } from './ServicesComp';
+
+
+export const AllPages = () => {
+    return (
+        <div>
+            <CampaignComp/>
+            <SolutionComp/>
+            <ServiceComp/>
+        </div>
+    )
+}
