+++ front-end/src/Components/SolutionCenter/SolutionDiscussion.js
@@ -1,0 +1,7 @@
+export const SolutionDiscussion = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
