+++ front-end/src/Components/SolutionCenter/SolutionNeeds.js
@@ -1,0 +1,7 @@
+export const SolutionNeeds = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
