+++ front-end/src/Components/SolutionCenter/SolutionWorks.js
@@ -1,0 +1,25 @@
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+import Placeholder from 'react-bootstrap/Placeholder';
+import Badge from 'react-bootstrap/Badge';
+import Image from 'react-bootstrap/Image';
+
+export const SolutionWorks = () => {
+    return (
+        <div>
+            <Row>
+                <Col>
+                    <strong>Explanation</strong>
+                </Col>
+                <Col className="text-end">
+                    <strong>Translate:</strong> original
+                </Col>
+            </Row>
+            <Row>
+                <Col>
+                
+                </Col>
+            </Row>
+        </div>
+    )
+}
